feature/refactor-easy-doctrine
- Refactor EasyDoctrine
